+++ spec/vm_spec.rb
@@ -17,103 +17,113 @@
 
 require "spec"
 
-describe "Pennyworth rspec helper" do
-  it "can start the system under test" do
-    $vm = start_system(box: "opensuse131")
-  end
+describe VM do
+  let(:ssh_output) { "-rw-r--r-- 1 root root 642 Sep 27 22:06 /etc/hosts" }
 
-  it "can inject a file" do
-    File.open("/tmp/helper.sh", "w") do |file|
-      file.write("#! /bin/bash
+  describe "#run_command" do
+    it "calls ssh and returns the command's standard output" do
+      expect(Cheetah).to receive(:run).
+        and_return(ssh_output)
 
-echo 'first output line'
-echo 'first error line' >&2
-echo 'second output line'
-echo 'second error line' >&2
-")
+      system = VM.new(nil)
+      system.ip = "1.2.3.4"
+      output = system.run_command("ls", "-l", "/etc/hosts", {:stdout=>:capture})
+
+      expect(output).to eq (ssh_output)
     end
-    $vm.inject_file("/tmp/helper.sh", "./", mode: "+x")
-    File.delete("/tmp/helper.sh")
-  end
 
-  it "can extract files" do
-    $vm.extract_file("/etc/SuSE-release", "/tmp")
-    File.open("/tmp/SuSE-release", "rb") do |file|
-      $out = file.read
+    it "executes commands as given user" do
+      expect(Cheetah).to receive(:run).
+        with(
+          "ssh", "-o", "UserKnownHostsFile=/dev/null", "-o", "StrictHostKeyChecking=no",
+          "root@1.2.3.4", "LC_ALL=C", "su", "-l", "vagrant", "-c", "ls", "-l", "/etc/hosts",
+          {:stdout=>:capture}).
+        and_return(ssh_output)
+
+      system = VM.new(nil)
+      system.ip = "1.2.3.4"
+      output = system.run_command("ls", "-l", "/etc/hosts", {as: "vagrant", stdout: :capture})
+
+      expect(output).to eq (ssh_output)
     end
-    expect($out).to start_with("openSUSE 13.1 (x86_64)")
-    File.delete("/tmp/SuSE-release")
-  end
 
-  it "can inject directories" do
-# TBD
-  end
+    it "raises ExecutionFailed in case of errors" do
+      expect(Cheetah).to receive(:run).and_raise(Cheetah::ExecutionFailed.new(nil, nil, nil, nil))
 
-  it "can capture output to a variable" do
-    $out = $vm.run_command("ls -l /", stdout: :capture)
-    expect($out).to match(/ etc$/)
-  end
-
-  it "can capture errors to a variable" do
-    $err = $vm.run_command("./helper.sh", stderr: :capture)
-    expect($err).to include("first error line\nsecond error line\n")
-  end
-
-  it "can capture output and errors to a variable" do
-    $out, $err = $vm.run_command("./helper.sh", stdout: :capture, stderr: :capture)
-    expect($out).to eq("first output line\nsecond output line\n")
-    expect($err).to include("first error line\nsecond error line\n")
-  end
-
-  it "does escape shell variables" do
-    $out = $vm.run_command("echo -n $HOME", stdout: :capture)
-    # '$HOME' and not '/home/testuser'!
-    expect($out).to eq("$HOME")
-  end
-
-  it "can capture output to a file" do
-    File.open("/tmp/out_test.txt", "w") do |file|
-      $vm.run_command("./helper.sh", stdout: file)
-    end
-    File.open("/tmp/out_test.txt", "rb") do |file|
-      expect(file.read).to eq("first output line\nsecond output line\n")
-    end
-    File.delete("/tmp/out_test.txt")
-  end
-
-  it "can capture errors to a file" do
-    File.open("/tmp/err_test.txt", "w") do |file|
-      $vm.run_command("./helper.sh", stderr: file)
-    end
-    File.open("/tmp/err_test.txt", "rb") do |file|
-      expect(file.read).to end_with("first error line\nsecond error line\n")
-    end
-    File.delete("/tmp/err_test.txt")
-  end
-
-  it "raises exceptions on errors" do
-    begin
-      $vm.run_command("ls -l /bang")
-      $out = "We should not reach this point."
-    rescue ExecutionFailed => e
-      $out = e.message
-    end
-    expect($out).to end_with("ls: cannot access /bang: No such file or directory\n\n")
-  end
-
-#  it "can take input from a string variable" do
-#    $out = $vm.run_command("cat", stdin: "Hello\nWorld!\n", stdout: :capture)
-#    expect($out).to eq("Hello\nWorld!\n")
-#  end
-
-  it "can take input from a file" do
-    File.open("/etc/passwd", "r") do |pw_file|
-      $out = $vm.run_command("cat", stdin: pw_file, stdout: :capture)
-      expect($out).to match(/^root:x:0:0:root/)
+      expect {
+        system = VM.new(nil)
+        system.ip = "1.2.3.4"
+        system.run_command("foo")
+      }.to raise_error(ExecutionFailed)
     end
   end
-end
 
-RSpec.configure do |config|
-  config.vagrant_dir = File.expand_path("examples/vagrant")
+  describe "#extract_file" do
+    it "calls scp with source and destination as arguments" do
+      expect(Cheetah).to receive(:run)
+
+      system = VM.new(nil)
+      system.ip = "1.2.3.4"
+      system.extract_file("/etc/hosts", "/tmp")
+    end
+  end
+
+  describe "#inject_file" do
+    it "calls scp with source and destination as arguments" do
+      expect(Cheetah).to receive(:run)
+
+      system = VM.new(nil)
+      system.ip = "1.2.3.4"
+      system.inject_file("/tmp/hosts", "/etc")
+    end
+
+    it "copies the file and sets the owner" do
+      expect(Cheetah).to receive(:run) { |*args| expect(args).to include(/scp/) }
+      expect(Cheetah).to receive(:run) { |*args| expect(args).to include(/chown.*tux/) }
+
+      system = VM.new(nil)
+      system.ip = "1.2.3.4"
+      system.inject_file("/tmp/hosts", "/etc", owner: "tux")
+    end
+
+    it "copies the file and sets the group" do
+      expect(Cheetah).to receive(:run) { |*args| expect(args).to include(/scp/) }
+      expect(Cheetah).to receive(:run) { |*args| expect(args).to include(/chown.*:tux/) }
+
+      system = VM.new(nil)
+      system.ip = "1.2.3.4"
+      system.inject_file("/tmp/hosts", "/etc", group: "tux")
+    end
+
+    it "copies the file and sets the mode" do
+      expect(Cheetah).to receive(:run) { |*args| expect(args).to include(/scp/) }
+      expect(Cheetah).to receive(:run) { |*args| expect(args).to include(/chmod 600/) }
+
+      system = VM.new(nil)
+      system.ip = "1.2.3.4"
+      system.inject_file("/tmp/hosts", "/etc", mode: "600")
+    end
+  end
+
+  describe "#inject_directory" do
+    it "calls scp with source and destination as arguments" do
+      expect(Cheetah).to receive(:run) { |*args| expect(args).to include(/mkdir -p/) }
+      expect(Cheetah).to receive(:run) { |*args| expect(args).to include(/scp/) }
+
+      system = VM.new(nil)
+      system.ip = "1.2.3.4"
+      system.inject_directory("/tmp/hosts", "/etc")
+    end
+
+    it "copies the directory and sets the user and group" do
+      expect(Cheetah).to receive(:run) { |*args| expect(args).to include(/mkdir -p/) }
+      expect(Cheetah).to receive(:run) { |*args| expect(args).to include(/scp/) }
+      expect(Cheetah).to receive(:run) { |*args| expect(args).to include(/chown -R user:group/) }
+
+      system = VM.new(nil)
+      system.ip = "1.2.3.4"
+      system.inject_directory("/tmp/hosts", "/etc", owner: "user", group: "group")
+    end
+  end
+
 end
